Revert "Move constants to `const`"
This reverts commit a7dd5c5b714ffd97782e3033f4eda8cfcd635386.
